requirements.txt
# conda env
streamlit
langchain_core
langchain_community
langchain_ollama
pdfplumber
langchain-ollama
google-genai
python-dotenv
sentence_transformers
google-generativeai
